+++ tests/ecs/react.rs
@@ -164,6 +164,11 @@ fn on_resource_mutation(mut rcommands: ReactCommands) -> RevokeToken
     rcommands.on(resource_mutation::<TestReactRes>(), update_test_recorder_with_resource)
 }
 
+fn on_resource_mutation_once(mut rcommands: ReactCommands) -> RevokeToken
+{
+    rcommands.once(resource_mutation::<TestReactRes>(), update_test_recorder_with_resource).unwrap()
+}
+
 fn on_event(mut rcommands: ReactCommands) -> RevokeToken
 {
     rcommands.on(event::<IntEvent>(), update_test_recorder_with_event)
@@ -615,6 +620,31 @@ fn react_resource_mutation()
 //-------------------------------------------------------------------------------------------------------------------
 
 #[test]
+fn react_resource_mutation_once()
+{
+    // setup
+    let mut app = App::new();
+    app.add_plugins(ReactPlugin)
+        .insert_react_resource(TestReactRes::default())
+        .init_resource::<TestReactRecorder>();
+    let mut world = &mut app.world;
+
+    // add reactor
+    syscall(&mut world, (), on_resource_mutation_once);
+    assert_eq!(world.resource::<TestReactRecorder>().0, 0);
+
+    // update resource (reaction)
+    syscall(&mut world, 100, update_react_res);
+    assert_eq!(world.resource::<TestReactRecorder>().0, 100);
+
+    // update resource (no reaction)
+    syscall(&mut world, 1, update_react_res);
+    assert_eq!(world.resource::<TestReactRecorder>().0, 100);
+}
+
+//-------------------------------------------------------------------------------------------------------------------
+
+#[test]
 fn react_event()
 {
     // prepare tracing
@@ -908,6 +938,30 @@ fn revoke_multiple_reactors()
     // mutate resource (no reaction)
     syscall(&mut world, 1, update_react_res);
     assert_eq!(world.resource::<TestReactRecorder>().0, 223);
+}
+
+//-------------------------------------------------------------------------------------------------------------------
+
+#[test]
+fn revoke_once_reactor()
+{
+    // setup
+    let mut app = App::new();
+    app.add_plugins(ReactPlugin)
+        .insert_react_resource(TestReactRes::default())
+        .init_resource::<TestReactRecorder>();
+    let mut world = &mut app.world;
+
+    // add reactor
+    let revoke_token = syscall(&mut world, (), on_resource_mutation_once);
+    assert_eq!(world.resource::<TestReactRecorder>().0, 0);
+
+    // revoke reactor
+    syscall(&mut world, revoke_token, revoke_reactor);
+
+    // mutate resource (no reaction)
+    syscall(&mut world, 1, update_react_res);
+    assert_eq!(world.resource::<TestReactRecorder>().0, 0);
 }
 
 //-------------------------------------------------------------------------------------------------------------------